# Conflicts:
#	src/app/header/header.component.ts

+++ src/app/forms/form.service.ts
@@ -18,7 +18,7 @@ export class FormService {
     private afs: AngularFirestore
   ) {
     this.pendingAdoptionFormsCollection = this.afs.collection('adoption-forms', ref => {
-      return ref.where('accepted', '==', false);
+      return ref.where('accepted', '==', false).where('rejected', '==', false);
     });
     this.adoptionForms = this.pendingAdoptionFormsCollection.snapshotChanges().pipe(map(changes => {
       return changes.map(a => {
@@ -46,9 +46,39 @@ export class FormService {
     return formRef.set(acceptedForm, {merge: true});
   }
 
-  rejectAdoptionForm(form: AdoptionForm): void{
+  rejectAdoptionForm(form: AdoptionForm): Promise<void>{
     const formRef: AngularFirestoreDocument<AdoptionForm> = this.afs.doc(`adoption-forms/${form.idForm}`);
-    formRef.delete();
+    const acceptedForm: AdoptionForm = {
+      ...form,
+      rejected: true,
+    };
+    return formRef.set(acceptedForm, {merge: true});
+  }
+
+  getUserAdoptionForms(userUid: string): Observable<AdoptionForm[]> {
+    const userAdoptionForms: AngularFirestoreCollection = this.afs.collection('adoption-forms', ref => {
+      return ref.where('userUid', '==', userUid).where('userReview', '==', true);
+    });
+    return userAdoptionForms.snapshotChanges().pipe(map(changes => {
+      return changes.map(a => {
+        const data = a.payload.doc.data() as AdoptionForm;
+        data.idForm = a.payload.doc.id;
+        return data;
+      });
+    }));
+  }
+
+  clearUserAdoptionForm(form: AdoptionForm): Promise<void>{
+    const formRef: AngularFirestoreDocument<AdoptionForm> = this.afs.doc(`adoption-forms/${form.idForm}`);
+    if (form.accepted){
+      const acceptedForm: AdoptionForm = {
+        ...form,
+        userReview: false
+      };
+      return formRef.set(acceptedForm, {merge : true});
+    } else {
+      return formRef.delete();
+    }
   }
 
 }
